Nominatim: move from `geopy.geocoders.osm` to `geopy.geocoders.nominatim`
